+++ addon/utils/_private/ember-node.ts
@@ -184,6 +184,7 @@ export function createEmberNodeSpec(config: EmberNodeConfig): NodeSpec {
     attrs,
     parseDOM,
     toDOM,
+    ...passthrough
   } = config;
   return {
     inline,
@@ -239,6 +240,7 @@ export function createEmberNodeSpec(config: EmberNodeConfig): NodeSpec {
             : [[inline ? 'span' : 'div', { 'data-slot': 'true' }, 0]]),
         ];
       }),
+    ...passthrough,
   };
 }