+++ tests/dummy/app/controllers/index.ts
@@ -47,6 +47,7 @@ import {
 } from '@lblod/ember-rdfa-editor/plugins/link';
 import { inject as service } from '@ember/service';
 import IntlService from 'ember-intl/services/intl';
+import { firefoxCursorFix } from '@lblod/ember-rdfa-editor/plugins/firefox-cursor-fix';
 
 export default class IndexController extends Controller {
   @tracked rdfaEditor?: SayController;
@@ -97,6 +98,7 @@ export default class IndexController extends Controller {
   }
 
   @tracked plugins: Plugin[] = [
+    firefoxCursorFix(),
     tablePlugin,
     tableKeymap,
     linkPasteHandler(this.schema.nodes.link),