+++ CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Add padding to rdfa-toggle
 - Fix shrinking issue with toolbar groups
+- Fix disappearing cursor problems in firefox
 
 ### Deprecated
 - Usage of `Plugins::List::IndentationControls`, use `Plugins::Indentation::IndentationMenu` instead.