+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.3.0] - 2023-03-02
 ### Added
 - Addition of a formatting toggle component which allows showing formatting marks
 - Addition of linkPasteHandler plugin which detects and converts links in pasted content
@@ -1477,7 +1479,8 @@ If you bump to this release, also upgrade ember-appuniversum to 1.0.0 along with
 
 
 
-[unreleased]: https://github.com/lblod/ember-rdfa-editor/compare/v3.2.1...HEAD
+[unreleased]: https://github.com/lblod/ember-rdfa-editor/compare/v3.3.0...HEAD
+[3.3.0]: https://github.com/lblod/ember-rdfa-editor/compare/v3.2.1...v3.3.0
 [3.2.1]: https://github.com/lblod/ember-rdfa-editor/compare/v3.2.0...v3.2.1
 [3.2.0]: https://github.com/lblod/ember-rdfa-editor/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/lblod/ember-rdfa-editor/compare/v3.0.0...v3.1.0